// frontend/src/codegen-pipeline/model/reproCheckEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ReproCheckEntryBuildCommit } from "./reproCheckEntryBuildCommit"
import type { ReproCheckEntryFinishedAt } from "./reproCheckEntryFinishedAt"
import type { ReproCheckEntryGitRepo } from "./reproCheckEntryGitRepo"
import type { ReproCheckEntryReprocheckLogUrl } from "./reproCheckEntryReprocheckLogUrl"
import type { ReproCheckEntryReprocheckStatus } from "./reproCheckEntryReprocheckStatus"
import type { ReproCheckEntryReproPipelineId } from "./reproCheckEntryReproPipelineId"
import type { ReproCheckEntryResultUrl } from "./reproCheckEntryResultUrl"

export interface ReproCheckEntry {
  app_id: string
  reprocheck_status: ReproCheckEntryReprocheckStatus
  result_url: ReproCheckEntryResultUrl
  build_commit: ReproCheckEntryBuildCommit
  git_repo: ReproCheckEntryGitRepo
  finished_at: ReproCheckEntryFinishedAt
  reprocheck_log_url: ReproCheckEntryReprocheckLogUrl
  repro_pipeline_id: ReproCheckEntryReproPipelineId
}