// frontend/src/codegen-pipeline/model/reproducibilityData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ReproCheckEntry } from "./reproCheckEntry"

export interface ReproducibilityData {
  reproducible: ReproCheckEntry[]
  unreproducible: ReproCheckEntry[]
  failed_to_rebuild: ReproCheckEntry[]
  unknown: ReproCheckEntry[]
}