// frontend/src/codegen-pipeline/model/reproducibleApiApiReproducibleGetStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

export type ReproducibleApiApiReproducibleGetStatus =
  | (typeof ReproducibleApiApiReproducibleGetStatus)[keyof typeof ReproducibleApiApiReproducibleGetStatus]
  | null

export const ReproducibleApiApiReproducibleGetStatus = {
  reproducible: "reproducible",
  unreproducible: "unreproducible",
  failed: "failed",
  none: "none",
} as const